Sanitize the back URL used when rendering PDF
